requirements.txt
altair
numpy
pandas
scikit-learn
scipy
streamlit
umap-learn